requirements.txt
gradio
torch
torch
pillow